requirements.txt
Django==1.6.4
MySQL-python==1.2.5
#PIL==1.1.7
Pillow==2.5.1
South==1.0
distribute==0.6.10
django-admin-jqueryui==1.8.22
django-ckeditor-updated==4.4.0
django-classy-tags==0.5.1
django-cms==3.0
django-inline-ordering==1.0.2
django-modeltranslation==0.8b2
django-mptt==0.6.0
django-recaptcha==1.0
django-sekizai==0.7
djangocms-admin-style==0.2.2
html5lib==1.0b3
six==1.7.3
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
#django-import-export==0.2.3
